mistakes are fixed

+++ src/components/App.js
@@ -15,7 +15,6 @@ import {CurrentUserContext} from '../contexts/CurrentUserContext';
 function App() {
 
     const [currentUser, setCurrentUser] = React.useState({});
-    const [currentPlace, setCurrentPlace] = React.useState({});
     const [cards, setCards] = React.useState([]);
 
     React.useEffect(() => {
@@ -66,37 +65,49 @@ function App() {
     function handleCardLike(card) {
         // Снова проверяем, есть ли уже лайк на этой карточке
         const isLiked = card.likes.some(i => i._id === currentUser._id);
+        if (isLiked) {
+            api.deleteLike(card._id, isLiked).then((newCard) => {
+                // Формируем новый массив на основе имеющегося, подставляя в него новую карточку
+                const newCards = cards.map((c) => c._id === card._id ? newCard : c);
+                // Обновляем стейт
+                setCards(newCards)
 
+            })
+                .catch((err) => {
+                    console.log(err);
+                });
+        } else {
+            api.putLike(card._id, !isLiked).then((newCard) => {
+                // Формируем новый массив на основе имеющегося, подставляя в него новую карточку
+                const newCards = cards.map((c) => c._id === card._id ? newCard : c);
+                // Обновляем стейт
+                setCards(newCards)
+
+            })
+                .catch((err) => {
+                    console.log(err);
+                });
+        }
         // Отправляем запрос в API и получаем обновлённые данные карточки
-        api.putLike(card._id, !isLiked).then((newCard) => {
-            // Формируем новый массив на основе имеющегося, подставляя в него новую карточку
-            const newCards = cards.map((c) => c._id === card._id ? newCard : c);
-            // Обновляем стейт
-            setCards(newCards);
-        });
-        api.deleteLike(card._id, isLiked).then((newCard) => {
-            // Формируем новый массив на основе имеющегося, подставляя в него новую карточку
-            const newCards = cards.map((c) => c._id === card._id ? newCard : c);
-            // Обновляем стейт
-            setCards(newCards);
-        });
     }
 
 
     function handleCardDelete(card) {
-
         // Отправляем запрос в API и получаем обновлённые данные карточки
         api.deleteCard(card._id).then(() => {
             // Формируем новый массив на основе имеющегося, подставляя в него новую карточку
             const newCards = cards.filter((c) => c._id !== card._id);
             // Обновляем стейт
-            setCards(newCards);
+            setCards(newCards)
+
+        }).catch((err) => {
+            console.log(err);
         });
     }
 
     function handleUpdateUser(data) {
-        api.updateUserInfo(data.name, data.about).then((r) => {
-            setCurrentUser(r);
+        api.updateUserInfo(data.name, data.about).then((updatedUserData) => {
+            setCurrentUser(updatedUserData);
             closeAllPopups()
         })
             .catch((err) => {
@@ -105,9 +116,8 @@ function App() {
     }
 
     function handleUpdatePlace(data) {
-        api.addNewCard(data.name, data.link).then((r) => {
-            setCurrentPlace(r);
-            setCards([...cards, r]);
+        api.addNewCard(data.name, data.link).then((newCardData) => {
+            setCards([...cards, newCardData]);
             closeAllPopups()
         })
             .catch((err) => {
@@ -117,8 +127,8 @@ function App() {
 
 
     function handleUpdateAvatar(data) {
-        api.setUserAvatar(data).then((r) => {
-            setCurrentUser(r);
+        api.setUserAvatar(data).then((updatedUserData) => {
+            setCurrentUser(updatedUserData);
             closeAllPopups()
         })
             .catch((err) => {
@@ -128,31 +138,29 @@ function App() {
 
 
     return (
-        <>
 
-            <div className="page">
-                <CurrentUserContext.Provider value={currentUser}>
-                    <Header/>
+        <div className="page">
+            <CurrentUserContext.Provider value={currentUser}>
+                <Header/>
 
-                    <Main onEditProfile={handleEditProfileClick} onAddPlace={handleAddPlaceClick}
-                          onEditAvatar={handleEditAvatarClick} cards={cards}
-                          onCardLike={handleCardLike} onCardDelete={handleCardDelete} onCardClick={handleCardClick}
-                    />
+                <Main onEditProfile={handleEditProfileClick} onAddPlace={handleAddPlaceClick}
+                      onEditAvatar={handleEditAvatarClick} cards={cards}
+                      onCardLike={handleCardLike} onCardDelete={handleCardDelete} onCardClick={handleCardClick}
+                />
 
-                    <EditAvatarPopup isOpen={isEditAvatarPopupOpen} onClose={closeAllPopups}
-                                     onUpdateAvatar={handleUpdateAvatar}/>
-                    <EditProfilePopup isOpen={isEditProfilePopupOpen} onClose={closeAllPopups}
-                                      onUpdateUser={handleUpdateUser}/>
-                    <AddPlacePopup isOpen={isAddPlacePopupOpen} onClose={closeAllPopups}
-                                   onUpdatePlace={handleUpdatePlace}/>
+                <EditAvatarPopup isOpen={isEditAvatarPopupOpen} onClose={closeAllPopups}
+                                 onUpdateAvatar={handleUpdateAvatar}/>
+                <EditProfilePopup isOpen={isEditProfilePopupOpen} onClose={closeAllPopups}
+                                  onUpdateUser={handleUpdateUser}/>
+                <AddPlacePopup isOpen={isAddPlacePopupOpen} onClose={closeAllPopups}
+                               onUpdatePlace={handleUpdatePlace}/>
 
-                    <PopupWithForm name={'sure'} title={'Вы уверены?'} buttonText={'Да'}/>
-                    <ImagePopup isOpen={selectedCard} onClose={closeAllPopups} image={dataImage}/>
+                <PopupWithForm name={'sure'} title={'Вы уверены?'} buttonText={'Да'}/>
+                <ImagePopup isOpen={selectedCard} onClose={closeAllPopups} image={dataImage}/>
 
-                    <Footer/>
-                </CurrentUserContext.Provider>
-            </div>
-        </>
+                <Footer/>
+            </CurrentUserContext.Provider>
+        </div>
     );
 }
